chore(ui): rename `context2.tsx` -> `context.tsx`
