+++ invokeai/frontend/web/src/features/controlLayers/components/SimpleSession/context.tsx
